Refactoring, non-defined method removed
