--- src/main/java/org/infoshareacademy/javaholics/Event/EventService.java
@@ -1,5 +1,0 @@
-package org.infoshareacademy.javaholics.Event;
-
-public class EventService {
-
-    }
