--- src/main/java/org/infoshareacademy/javaholics/RouteService.java
@@ -1,4 +1,0 @@
-package src.main.java.org.infoshareacademy.javaholics;
-
-public class RouteService {
-}
