--- src/main/java/org/infoshareacademy/javaholics/User/UserService.java
@@ -1,4 +1,0 @@
-package org.infoshareacademy.javaholics.User;
-
-public class UserService {
-}
