+++ src/main/java/org/infoshareacademy/javaholics/event/EventService.java
@@ -1,0 +1,5 @@
+package org.infoshareacademy.javaholics.event;
+
+public class EventService {
+
+    }
