+++ src/main/java/org/infoshareacademy/javaholics/route/RouteService.java
@@ -1,0 +1,4 @@
+package org.infoshareacademy.javaholics.route;
+
+public class RouteService {
+}
