+++ src/main/java/org/infoshareacademy/javaholics/user/UserService.java
@@ -1,0 +1,4 @@
+package org.infoshareacademy.javaholics.user;
+
+public class UserService {
+}
